[General Master] Move MASTER_FORM_CONSTANTS to constants file
