+++ src/containers/MasterPage/components/MasterForm.jsx
@@ -7,6 +7,7 @@ import renderSelectField from '../../../shared/components/form/Select';
 import validate from './validate';
 import {renderField} from "../../../shared/components/form/InputField";
 import LoadingSpinner from "../../../shared/components/loading_spinner/LoadingSpinner";
+import MASTER_FORM_CONSTANTS from "../constants";
 
 class MasterForm extends Component {
 	static propTypes = {
@@ -482,50 +483,6 @@ class MasterForm extends Component {
 		);
 	}
 }
-
-export const MASTER_FORM_CONSTANTS = {
-	masterFormName: 'MasterForm',
-	field: {
-		masCd: {
-			name: 'masCd',
-			maxLength: 20,
-		},
-		hiddenMasCdDuplicatedChecker: 'hiddenMasCdDuplicatedChecker',
-		masCdNm: {
-			name: 'masCdName',
-			maxLength: 50,
-		},
-
-		catCdNm: 'cateCdName',
-		catCd: 'cateCd',
-
-		parentMasNm: 'parentMasName',
-		parentMasCd: 'parentMasCd',
-
-		virtualYn: 'virtualYn',
-		activeYn: 'activeYn',
-		sysCodeYn: 'sysCodeYn',
-
-		processingSeq: {
-			name: 'processingSeq',
-			maxLength: 6,
-		},
-
-		definitionValue: 'definitionValue',
-		temperature: 'temperature',
-		pressure: 'pressure',
-		curingTime: 'curingTime',
-
-		description: 'remark',
-	},
-	definitionValueRange: [0, 1, 2, 3, 4, 5, 6, 7, 8],
-	submissionState: {
-		failed: -1,
-		initial: 0,
-		onGoing: 1,
-		done: 2,
-	}
-};
 
 export default reduxForm({
 	form: MASTER_FORM_CONSTANTS.masterFormName,
